requirements.txt
fastapi
uvicorn
pydantic
pandas
pyarrow
databricks-sdk
databricks-sql-connector
requests
python-dotenv
pycloudflared
